+++ crates/vp_multitoken/src/keys.rs
@@ -1,5 +1,4 @@
 //! key helpers
-use std::str::FromStr;
 
 use anoma_vp_prelude::storage::{DbKeySeg, Key};
 use shared::keys::BALANCE_KEY_SEGMENT;
@@ -7,16 +6,12 @@ use shared::keys::BALANCE_KEY_SEGMENT;
 const MULTITOKEN_KEY_SEGMENT: &str = "ERC20";
 
 pub fn balance(multitoken_addr: &str, token_id: &str, owner_addr: &str) -> Key {
-    Key::from_str(multitoken_addr)
-        .unwrap()
-        .push(&MULTITOKEN_KEY_SEGMENT.to_owned())
-        .unwrap()
-        .push(&token_id.to_owned())
-        .unwrap()
-        .push(&BALANCE_KEY_SEGMENT.to_owned())
-        .unwrap()
-        .push(&owner_addr.to_owned())
-        .unwrap()
+    shared::keys::balance(
+        multitoken_addr,
+        MULTITOKEN_KEY_SEGMENT,
+        token_id,
+        owner_addr,
+    )
 }
 
 pub fn is_balance_key(key: &Key) -> bool {
@@ -28,21 +23,9 @@ pub fn is_balance_key(key: &Key) -> bool {
 
 #[cfg(test)]
 mod tests {
+    use std::str::FromStr;
+
     use super::*;
-
-    const MULTITOKEN_ADDRESS: &str =
-        "atest1v4ehgw36g4pyg3j9x3qnjd3cxgmyz3fk8qcrys3hxdp5xwfnx3zyxsj9xgunxsfjg5u5xvzyzrrqtn";
-    const ERC20_ADDRESS: &str = "0x6B175474E89094C44Da98b954EedeAC495271d0F";
-    const OWNER_ADDRESS: &str =
-        "atest1d9khqw36x9zyxwfhgfpygv2pgc65gse4gy6rjs34gfzr2v69gy6y23zpggurjv2yx5m52sesu6r4y4";
-
-    #[test]
-    fn test_balance() {
-        assert_eq!(
-                "atest1v4ehgw36g4pyg3j9x3qnjd3cxgmyz3fk8qcrys3hxdp5xwfnx3zyxsj9xgunxsfjg5u5xvzyzrrqtn/ERC20/0x6B175474E89094C44Da98b954EedeAC495271d0F/balance/atest1d9khqw36x9zyxwfhgfpygv2pgc65gse4gy6rjs34gfzr2v69gy6y23zpggurjv2yx5m52sesu6r4y4",
-                balance(MULTITOKEN_ADDRESS, ERC20_ADDRESS, OWNER_ADDRESS).to_string()
-            )
-    }
 
     #[test]
     fn test_is_balance_key() {
